refactor: re-organize components
